mysite_delete/requirements.txt
asgiref==3.5.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
charset-normalizer==2.1.1
Django==4.1.3
django-haystack==3.2.1
django-mdeditor==0.1.20
django-resized==1.0.2
django-tinymce==3.5.0
elasticsearch==7.17.7
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2-binary==2.9.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
zipp==3.10.0
